requirements.txt
psycopg2_binary==2.9.1
PyPika==0.48.9
streamlit==1.21.0
